- updated to use the new "DeployFolder" parameter from cake-config.json. The build will now send the output to Deployfolder\Site\HabitatHome and Deployfolder\Site\xConnect. It will also use the Deploy Folder for whatever else was it doing with the output folders previously.
 - Removed instanceURL from the cake-config.json due to us not using it and json not supporting commenting things out

As for passing values, we have 2 options, we can have some code check for existing files in each script. Or we can try to pass a variable containing the information of which files already exist at build start between the files. This can be done via setting a return value in the powershell script and catching it in a variable in the build file. We would then attempt to pass that variable as an argument to the next script and so on. Besides time there really isn't a huge issue with checking once each file.